requirements.txt
pyyaml
opera
